requirements.txt
# Core Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Streamlit Web App
streamlit>=1.25.0

# Image Processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Explainable AI (XAI)
grad-cam>=1.4.0
pytorch-grad-cam>=1.4.0

# Generative AI Integration
google-generativeai>=0.3.0

# Data Science & Utilities
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Google Cloud Storage (for Colab training)
# Uncomment if using GCS in Colab
# google-cloud-storage>=2.10.0

# Jupyter Notebook Support
# Uncomment for notebook development
# jupyter>=1.0.0
# ipywidgets>=8.0.0
